// apps/nk-vrapce/src/types/hns.ts
export interface HnsTeam {
  id: number | null;
  fifaId: string | null;
  name: string | null;
  picture: string | null;
  place: string | null;
  allowDetail: boolean | null;
}
// ...
export interface PastMatch {
  matchId: number | null;
  result: string | null;
}

export interface TeamRanking {
  team: HnsTeam | null;
  played: number | null;
  wins: number | null;
  draws: number | null;
  losses: number | null;
  winsAET: number | null;
  winsAP: number | null;
  lossesAP: number | null;
  goalsFor: number | null;
  goalsAgainst: number | null;
  points: number | null;
  negativePoints: number | null;
  position: number | null;
  highlight: boolean | null;
  m1: PastMatch | null;
  m2: PastMatch | null;
  m3: PastMatch | null;
  m4: PastMatch | null;
  m5: PastMatch | null;
}